Added test files and additional functions. still WIP as it needs cleanup.

+++ avt_vimba_camera/include/avt_vimba_camera/avt_vimba_camera.hpp
@@ -93,7 +93,7 @@ public:
 
   // Getters
   CameraState getCameraState() const;
-  double getTimestamp();
+  u_int64_t getTimestamp();
   double getDeviceTemp();
   int getPtpOffset();
   int getImageWidth();
